+++ src/main/java/io/github/monadrome/parallelinscope/cancel/package-info.java
@@ -1,5 +1,0 @@
-/** Cooperative cancellation tokens, checkpoints, states, and exceptions. */
-@ParametersAreNonnullByDefault
-package io.github.monadrome.parallelinscope.cancel;
-
-import javax.annotation.ParametersAreNonnullByDefault;