+++ src/main/java/io/github/monadrome/parallelinscope/context/graph/package-info.java
@@ -1,5 +1,0 @@
-/** Task and executor dependency graphs used for potential deadlock detection. */
-@ParametersAreNonnullByDefault
-package io.github.monadrome.parallelinscope.context.graph;
-
-import javax.annotation.ParametersAreNonnullByDefault;