+++ src/main/java/io/github/monadrome/parallelinscope/internal/package-info.java
@@ -1,5 +1,0 @@
-/** Internal scheduling, task wrapping, and future inspection implementation. */
-@ParametersAreNonnullByDefault
-package io.github.monadrome.parallelinscope.internal;
-
-import javax.annotation.ParametersAreNonnullByDefault;