+++ src/main/java/io/github/monadrome/parallelinscope/scope/BlockingRisk.java
@@ -1,9 +1,0 @@
-package io.github.monadrome.parallelinscope.scope;
-
-/** Conservative blocking capability of a supplied executor. */
-public enum BlockingRisk {
-    UNKNOWN,
-    BOUNDED_PLATFORM_POOL,
-    VIRTUAL_THREAD_PER_TASK,
-    UNBOUNDED
-}