+++ src/main/java/io/github/monadrome/parallelinscope/spi/package-info.java
@@ -1,5 +1,0 @@
-/** Service provider interfaces for executor resolution and runtime monitoring. */
-@ParametersAreNonnullByDefault
-package io.github.monadrome.parallelinscope.spi;
-
-import javax.annotation.ParametersAreNonnullByDefault;